+++ adapters/Fyber/FyberAdapter/GADMAdapterFyberUtils.h
@@ -39,7 +39,8 @@ IAAdRequest *_Nonnull GADMAdapterFyberBuildRequestWithSpotIDAndAdConfiguration(
 /// Creates an ad request object for Fyber from the specified information.
 IAAdRequest *_Nonnull GADMAdapterFyberBuildRequestWithSpotID(NSString *_Nonnull spotID,
                                                              NSString *_Nullable keywords,
-                                                             CLLocation *_Nullable location);
+                                                             CLLocation *_Nullable location,
+                                                             IAUserData *_Nullable userData);
 
 /// Initializes the Fyber SDK with the given app ID. Returns YES if Fyber is initialized
 /// successfully, otherwise NO and sets |error|.