+++ adapters/Fyber/FyberAdapter/GADMediationAdapterFyber.m
@@ -16,6 +16,7 @@
 #import "GADMAdapterFyberConstants.h"
 #import "GADMAdapterFyberRewardedAd.h"
 #import "GADMAdapterFyberUtils.h"
+#import "GADMAdapterFyberExtras.h"
 
 #import <CoreLocation/CoreLocation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
@@ -36,7 +37,7 @@
   /// Fyber fullscreen controller to catch interstitial related ad events.
   IAFullscreenUnitController *_fullscreenUnitController;
 
-  /// Fyber mraid controller to support HTML ads and to catch MRAID content related callbacks.
+  /// Fyber MRAID controller to support HTML ads and to catch MRAID content related callbacks.
   IAMRAIDContentController *_MRAIDContentController;
 
   /// Fyber video controller to support VAST ads and to catch video progress events.
@@ -102,7 +103,7 @@
 }
 
 + (Class<GADAdNetworkExtras>)networkExtrasClass {
-  return Nil;
+  return GADMAdapterFyberExtras.class;
 }
 
 - (instancetype)initWithGADMAdNetworkConnector:(id<GADMAdNetworkConnector>)connector {
@@ -141,9 +142,28 @@
     if (error) {
       GADMAdapterFyberLog(@"Failed to load banner ad: %@", error.localizedDescription);
       [strongConnector adapter:strongSelf didFailAd:error];
-    } else {
-      [strongConnector adapter:strongSelf didReceiveAdView:strongSelf->_viewUnitController.adView];
+      return;
     }
+
+    // Verify the loaded ad size with the requested ad size.
+    GADAdSize loadedAdSize =
+        GADAdSizeFromCGSize(CGSizeMake(strongSelf->_viewUnitController.adView.frame.size.width,
+                                       strongSelf->_viewUnitController.adView.frame.size.height));
+    NSArray<NSValue *> *potentials = @[ NSValueFromGADAdSize(loadedAdSize) ];
+    GADAdSize closestSize = GADClosestValidSizeForAdSizes(adSize, potentials);
+    if (!IsGADAdSizeValid(closestSize)) {
+      NSString *logMessage = [NSString
+          stringWithFormat:@"The loaded ad size did not match the requested ad size. Requested ad "
+                           @"size: %@. Loaded size: %@.",
+                           NSStringFromGADAdSize(adSize), NSStringFromGADAdSize(loadedAdSize)];
+      GADMAdapterFyberLog(@"Failed to load banner ad: %@", logMessage);
+      NSError *error =
+          GADMAdapterFyberErrorWithCodeAndDescription(kGADErrorMediationInvalidAdSize, logMessage);
+      [strongConnector adapter:strongSelf didFailAd:error];
+      return;
+    }
+
+    [strongConnector adapter:strongSelf didReceiveAdView:strongSelf->_viewUnitController.adView];
   }];
 }
 
@@ -196,11 +216,11 @@
 
 - (void)presentInterstitialFromRootViewController:(UIViewController *)rootViewController {
   if (_fullscreenUnitController.isPresented) {
-      GADMAdapterFyberLog(@"Failed to show interstitial ad, it is already presented");
+    GADMAdapterFyberLog(@"Failed to show interstitial ad, it is already presented");
   } else if (!_fullscreenUnitController.isReady) {
-      GADMAdapterFyberLog(@"Failed to show interstitial ad, it has already expired");
+    GADMAdapterFyberLog(@"Failed to show interstitial ad, it has already expired");
   } else {
-      [_fullscreenUnitController showAdAnimated:YES completion:nil];
+    [_fullscreenUnitController showAdAnimated:YES completion:nil];
   }
 }
 
@@ -212,7 +232,6 @@
       }];
 
   GADMediationAdapterFyber *__weak weakSelf = self;
-
   _viewUnitController =
       [IAViewUnitController build:^(id<IAViewUnitControllerBuilder> _Nonnull builder) {
         GADMediationAdapterFyber *strongSelf = weakSelf;
@@ -226,7 +245,7 @@
 
   _adSpot = [IAAdSpot build:^(id<IAAdSpotBuilder> _Nonnull builder) {
     builder.adRequest = request;
-    builder.mediationType = [IAMediationAdMob new];
+    builder.mediationType = [[IAMediationAdMob alloc] init];
 
     GADMediationAdapterFyber *strongSelf = weakSelf;
     if (!strongSelf) {
@@ -261,7 +280,7 @@
 
   _adSpot = [IAAdSpot build:^(id<IAAdSpotBuilder> _Nonnull builder) {
     builder.adRequest = request;
-    builder.mediationType = [IAMediationAdMob new];
+    builder.mediationType = [[IAMediationAdMob alloc] init];
 
     GADMediationAdapterFyber *strongSelf = weakSelf;
     if (!strongSelf) {