+++ adapters/MoPub/MoPubAdapter/GADMAdapterMoPub.m
@@ -1,24 +1,13 @@
 #import "GADMAdapterMoPub.h"
 
+#import <MoPubSDK/Internal/Utility/MPImageDownloadQueue.h>
+#import <MoPubSDK/Internal/Utility/MPNativeCache.h>
+#import <MoPubSDK/MoPub.h>
 #import "GADMAdapterMoPubConstants.h"
 #import "GADMAdapterMoPubSingleton.h"
 #import "GADMAdapterMoPubUtils.h"
 #import "GADMAdapterMopubUnifiedNativeAd.h"
 #import "GADMoPubNetworkExtras.h"
-#import "MPAdView.h"
-#import "MPImageDownloadQueue.h"
-#import "MPInterstitialAdController.h"
-#import "MPLogging.h"
-#import "MPNativeAd.h"
-#import "MPNativeAdConstants.h"
-#import "MPNativeAdDelegate.h"
-#import "MPNativeAdRequest.h"
-#import "MPNativeAdRequestTargeting.h"
-#import "MPNativeAdUtils.h"
-#import "MPNativeCache.h"
-#import "MPStaticNativeAdRenderer.h"
-#import "MPStaticNativeAdRendererSettings.h"
-#import "MoPub.h"
 
 static NSMapTable<NSString *, GADMAdapterMoPub *> *GADMAdapterMoPubInterstitialDelegates;