+++ adapters/MoPub/MoPubAdapter/GADMediationAdapterMoPub.m
@@ -1,12 +1,11 @@
 #import "GADMediationAdapterMoPub.h"
 
+#import <MoPubSDK/MoPub.h>
 #import "GADMAdapterMoPubConstants.h"
 #import "GADMAdapterMoPubSingleton.h"
 #import "GADMAdapterMoPubUtils.h"
 #import "GADMMoPubRewardedAd.h"
 #import "GADMoPubNetworkExtras.h"
-#import "MPRewardedVideo.h"
-#import "MoPub.h"
 
 @implementation GADMediationAdapterMoPub {
   /// MoPub rewarded ad wrapper.