+++ example/Adapter/SampleAdapterMediatedNativeAd.h
@@ -26,7 +26,7 @@
 /// by a mediated network don't always line up with the ones expected by the Google
 /// Mobile Ads SDK (one might have "title" while the other expects "headline," for
 /// example). It's the job of this "mapper" class to smooth out those wrinkles.
-@interface SampleAdapterMediatedNativeAd : NSObject<GADMediatedUnifiedNativeAd, GADMediatedNativeAd>
+@interface SampleAdapterMediatedNativeAd : NSObject <GADMediatedUnifiedNativeAd>
 
 - (null_unspecified instancetype)init NS_UNAVAILABLE;