+++ adapters/Maio/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## maio iOS Mediation Adapter Changelog
 
+#### Version 1.5.5.0
+- Verified compatibility with maio SDK version 1.5.5.
+- Now requires Google Mobile Ads SDK version 7.64.0 or higher.
+
+Build and tested with:
+- Google Mobile Ads SDK version 7.64.0.
+- Maio SDK version 1.5.5.
+
 #### Version 1.5.4.0
 - Verified compatibility with maio SDK version 1.5.4.
 - Now requires Google Mobile Ads SDK version 7.60.0 or higher.