+++ adapters/MoPub/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## MoPub iOS Mediation Adapter Changelog
 
+####  Version 5.13.1.1
+- Fixed a memory leak issue with native ads.
+- Now requires Google Mobile Ads SDK version 7.65.0 or higher.
+
+Built and tested with
+- Google Mobile Ads SDK version 7.65.0.
+- MoPub SDK version 5.13.1.
+
 ####  Version 5.13.1.0
 - Verified compatibility with MoPub SDK 5.13.1.
 - Now requires Google Mobile Ads SDK version 7.61.0 or higher.