+++ adapters/Vungle/CHANGELOG.md
@@ -1,7 +1,21 @@
 ## Vungle iOS Mediation Adapter Changelog
 
-#### Next Version
+#### Version 6.8.0.0
+- Verified compatibility with Vungle SDK 6.8.0.
+- Now requires Google Mobile Ads SDK version 7.65.0 or higher.
+
+Build and tested with
+- Google Mobile Ads SDK version 7.65.0.
+- Vungle SDK version 6.8.0.
+
+#### Version 6.7.1.0
+- Verified compatibility with Vungle SDK 6.7.1.
+- Now requires Google Mobile Ads SDK version 7.64.0 or higher.
 - Fixed an issue where `didFailToPresentWithError:` was not called when a rewarded ad failed to present.
+
+Build and tested with
+- Google Mobile Ads SDK version 7.64.0.
+- Vungle SDK version 6.7.1.
 
 #### Version 6.7.0.0
 - Verified compatibility with Vungle SDK 6.7.0.